+++ .swiftlint.yml
@@ -1,0 +1,1 @@
+Sources/.swiftlint.yml
