+++ Package.swift
@@ -18,7 +18,11 @@ let package = Package(
     .package(url: "https://github.com/unrelentingtech/SwiftCBOR", from: "0.4.3"),
     .package(url: "https://github.com/SwiftyJSON/SwiftyJSON", from: "5.0.1"),
     .package(url: "https://github.com/Alamofire/Alamofire", from: "5.4.3"),
-    .package(name: "JSONSchema", url: "https://github.com/eu-digital-green-certificates/JSONSchema.swift", .branch("master"))
+    .package(
+      name: "JSONSchema",
+      url: "https://github.com/eu-digital-green-certificates/JSONSchema.swift",
+      .branch("master")
+    )
   ],
   targets: [
     // Targets are the basic building blocks of a package. A target can define a module or a test suite.
